refresh Token4

+++ src/HttpService.js
@@ -7,6 +7,9 @@ class HttpService {
         this.authStore = rootStore.authStore;
 
         this.clientID = 'q3lcLkv9AKycjIR7EQ8DMwAWdDaFDipiT78CiFvx'
+        this.refreshSubscribers = [];
+        this.isRefreshingToken = false;
+
         axios.defaults.baseURL = 'http://localhost:8003'
         axios.defaults.headers.common['Authorization'] = this.authStore.authToken
         reaction(()=>this.authStore.authToken, () =>{
@@ -23,17 +26,40 @@ class HttpService {
                     alert('로그인이 필요한 서비스입니다.');
                     this.rootStore.history.push('/login');
                 } else {
-                    return new Promise((resolve, reject) => {
-                        this.refreshToken().then(token => {
-                            originalRequest.headers.Authorization = this.authStore.authToken
-                            resolve(axios(originalRequest));
-                        }).catch(error => {
-                            this.authStore.deleteToken();
-                            reject(originalError);
-                            alert('로그인이 필요한 서비스입니다.');
-                            this.rootStore.history.push('/login');
+                    if (!this.isRefreshingToken) {
+                        this.isRefreshingToken = true;
+                        return new Promise((resolve, reject) => {
+                            this.refreshToken().then(token => {
+                                originalRequest.headers.Authorization = this.authStore.authToken
+                                resolve(axios(originalRequest));
+                                for(let subscriber of this.refreshSubscribers){
+                                    subscriber(token);
+                                }
+                            }).catch(error => {
+                                this.authStore.deleteToken();
+                                reject(originalError);
+                                alert('로그인이 필요한 서비스입니다.');
+                                this.rootStore.history.push('/login');
+                                for(let subscriber of this.refreshSubscribers){
+                                    subscriber(null);
+                                }
+                            }).finally(()=>{
+                                this.refreshSubscribers = [];
+                                this.isRefreshingToken = false;
+                            })
+                        });
+                    }
+                    return new Promise((resolve, reject) =>{
+                        this.refreshSubscribers.push((token) =>{
+                            if(token ==null){
+                                reject(originalError);
+                            }else{
+                                originalRequest.headers.Authorization = this.authStore.authToken
+                                resolve(axios(originalRequest));
+                            }
+                            resolve(123);
                         })
-                    });
+                    })
                 }
             }
             return Promise.reject(originalError);
@@ -50,6 +76,7 @@ class HttpService {
     getMe() {
         return axios.get('/me/')
         .then((response) => {
+            console.log(response);
             return response.data
         });
     }
@@ -57,6 +84,7 @@ class HttpService {
     indexMyItems(){
         return axios.get('/me/items/')
         .then((response) => {
+            console.log(response);
             return response.data
         });
     }
